+++ CHANGELOG.md
@@ -7,7 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - BREAKING: `get_recommended_labels` and `with_recommended_labels` now takes a struct of named arguments ([#501]).
+- BREAKING: `get_recommended_labels` (and co) now takes the operator and controller names separately ([#492]).
 
+[#492]: https://github.com/stackabletech/operator-rs/pull/492
 [#501]: https://github.com/stackabletech/operator-rs/pull/501
 
 ## [0.26.1] - 2022-11-08