+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - BREAKING: `get_recommended_labels` and `with_recommended_labels` now takes a struct of named arguments ([#501]).
 - BREAKING: `get_recommended_labels` (and co) now takes the operator and controller names separately ([#492]).
+- BREAKING: `ClusterResources` now takes the operator and controller names separately ([#492]).
+  - When upgrading, please use FQDN-style names for the operators (`{operator}.stackable.tech`).
 
 [#492]: https://github.com/stackabletech/operator-rs/pull/492
 [#501]: https://github.com/stackabletech/operator-rs/pull/501
@@ -33,12 +35,9 @@ All notable changes to this project will be documented in this file.
 - kube-rs: 0.74.0 -> 0.75.0 ([#490]).
 - BREAKING: `Client` methods now take the namespace as a `&str` (for namespaced resources) or
   `&()` (for cluster-scoped resources), rather than always taking an `Option<&str>` ([#490]).
-- BREAKING: `ClusterResources` and `with_recommended_labels` now take separate arguments for the operator and controller names ([#492]).
-  - When upgrading, please use FQDN-style names for the operators (`{operator}.stackable.tech`).
 
 [#445]: https://github.com/stackabletech/operator-rs/pull/445
 [#490]: https://github.com/stackabletech/operator-rs/pull/490
-[#492]: https://github.com/stackabletech/operator-rs/pull/492
 
 ## [0.25.3] - 2022-10-13