+++ CHANGELOG.md
@@ -7,8 +7,10 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - BREAKING: `get_recommended_labels` and `with_recommended_labels` now takes a struct of named arguments ([#501]).
+- Bump opentelemetry crates ([#502]).
 
 [#501]: https://github.com/stackabletech/operator-rs/pull/501
+[#502]: https://github.com/stackabletech/operator-rs/pull/502
 
 ## [0.26.1] - 2022-11-08