+++ CHANGELOG.md
@@ -7,9 +7,11 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - BREAKING: `get_recommended_labels` and `with_recommended_labels` now takes a struct of named arguments ([#501]).
+- Bump opentelemetry crates ([#502]).
 - Bump clap to 4.0 ([#503]).
 
 [#501]: https://github.com/stackabletech/operator-rs/pull/501
+[#502]: https://github.com/stackabletech/operator-rs/pull/502
 [#503]: https://github.com/stackabletech/operator-rs/pull/503
 
 ## [0.26.1] - 2022-11-08